requirements.txt
reflex>=0.7.11
openai>=1.78.1
ollama>=0.1.0
google-genai>=1.0.0
python-dotenv>=1.0.0
